% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_structure.R
\name{agify}
\alias{agify}
\title{Agify function}
\usage{
agify(name, country_id = NULL, simplify = TRUE, apikey = NULL, meta = FALSE)
}
\arguments{
\item{name}{Name/s to evaluate from. Can be a single string or a vector.}

\item{country_id}{Optional parameter to estimate for a specific country. ISO2 code countrycode.}

\item{simplify}{Defines if the result should be returned as a single vector or a \code{data.frame} with additional information.
By default set to \code{TRUE}, which returns a vector.}

\item{apikey}{API KEY from \href{https://agify.io/}{Agify.io}.}

\item{meta}{Logical parameter to define if API related information should be returned. By default set to \code{FALSE}.}
}
\value{
The estimated nationality in a single character vector form or a \code{data.frame} with additional information (e.g. probability).
}
\description{
Returns the estimated age from a given name.
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
