% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemografixeR.R
\docType{package}
\name{DemografixeR}
\alias{DemografixeR}
\title{DemografixeR: Estimate gender, age and nationality from a name}
\description{
The DemografixeR is an API wrapper for the genderize.io, agify.io &
nationalize.io APIs. Through the APIs it retrieves the estimated gender,
estimated age and/or estimated nationality from a name. The package most
important functions are the following:
 \itemize{
 \item \code{\link{genderize}}
 \item \code{\link{agify}}
 \item \code{\link{nationalize}}
 }
}
