% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_structure.R
\name{get_key}
\alias{get_key}
\title{Get previously saved API keys}
\usage{
get_key(type)
}
\arguments{
\item{type}{Obligatory parameter to define which key to retrieve:
\itemize{
\item \code{genderize} - Genderize.io API key
\item \code{agify} - Agify.io API key
\item \code{nationalize} - Nationalize.io API key
}}
}
\value{
Returns the saved API key in the environment. If no API key has been saved, returns \code{NULL} value.
}
\description{
Function to get the previously saved API key. It has the advantage that it is not necessary to explicitly publish the key in the users code.
Just do it one time and you're set. To save an API, use the \code{save_key()} function.
To remove a previously saved key, use the \code{remove_key()} function.
}
\section{Warning}{

Please be careful when dealing with API keys and other secrets & tokens - keep them private and do not publish them.
}

\examples{
\dontrun{
get_key(type="genderize")
get_key(type="agify")
get_key(type="nationalize")
}
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
