% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_structure.R
\name{remove_api_key}
\alias{remove_api_key}
\title{Removes saved API key}
\usage{
remove_api_key(verbose = TRUE)
}
\arguments{
\item{verbose}{\code{Logical} parameter to define if a verbose message should
be printed. By default set to \code{TRUE}.}
}
\value{
Does not return any object.
}
\description{
Removes saved API keys for the DemografixeR APIs (Genderize.io,
  Agify.io, Nationalize.io).
}
\section{Warning}{
 Please be careful when dealing with API keys and other
  secrets & tokens - keep them private and do not publish them.
}

\examples{
\dontrun{
remove_api_key()
}
}
\seealso{
\code{\link{save_api_key}}, \code{\link{get_api_key}}
}
\author{
Matthias Brenninkmeijer -
  \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
