% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_structure.R
\name{remove_key}
\alias{remove_key}
\title{Removes saved key}
\usage{
remove_key(type, verbose = TRUE)
}
\arguments{
\item{type}{Choose the type of key to remove from the environment variables:
\itemize{
\item \code{gender} - Genderize.io key
\item \code{age} - Agify.io key
\item \code{nationality} - Nationalize.io key
}}

\item{verbose}{Logical parameter to define if a verbose message should be shown.}
}
\value{
Does not return any object
}
\description{
Removes saved keys for the DemografixeR APIs (Genderize.io, Agify.io, Nationalize.io)
}
\examples{
\dontrun{
remove_key(type="gender")
remove_key(type="age")
remove_key(type="nationality")
}
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
