% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_structure.R
\name{nationalize}
\alias{nationalize}
\title{Nationalize function}
\usage{
nationalize(name, simplify = TRUE, sliced = TRUE, apikey = NULL, meta = FALSE)
}
\arguments{
\item{name}{Name/s to evaluate from. Can be a single string or a vector.}

\item{simplify}{Defines if the result should be returned as a single vector or a \code{data.frame} with additional information.
By default set to \code{TRUE}, which returns a vector.}

\item{sliced}{Names can have multiple estimated countries ranked by probabilty. This logical parameter allows to "slice"/keep only
the parameter with the highest probability to keep a single estimate for each name.}

\item{apikey}{API KEY from \href{https://nationalize.io/}{Nationalize.io}.}

\item{meta}{Logical parameter to define if API related information should be returned. By default set to \code{FALSE}.}
}
\value{
The estimated nationality in a single character vector form or a \code{data.frame} with additional information (e.g. probability).
}
\description{
Returns the estimated nationality from a given name.
}
\author{
Matthias Brenninkmeijer - \href{https://github.com/matbmeijer}{https://github.com/matbmeijer}
}
